Cleanup the mouse handling. Platform drivers should only return an integer with at least 16 bits of position information, the rest is handled in mouse.c and ps2mouse.c as required by the type of mouse, like speed limiting. Now at least on Linux I don't see pointer jumps anymore, except on very high speed. Straight movement is straight now that the rate limiting was fixed. Overall it's quite ok here now, let's hope the other platform are not broken by mistake.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@25525 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/os2/mousedrv.h
@@ -36,8 +36,8 @@ extern void mousedrv_init(void);
 
 extern void mousedrv_mouse_changed(void);
 
-extern BYTE mousedrv_get_x(void);
-extern BYTE mousedrv_get_y(void);
+extern int mousedrv_get_x(void);
+extern int mousedrv_get_y(void);
 extern unsigned long mousedrv_get_timestamp(void);
 
 /* ------------------ OS/2 specific ------------------- */
